feat: cdist compute_mode compatibility
